committed

+++ src/main/java/com/sunglowsys/repository/HotelRoomTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HotelRoomTypeRepository extends JpaRepository<HotelRoomType, Integer> {
+public interface HotelRoomTypeRepository extends JpaRepository<HotelRoomType, Long> {
 }
